refactor: Split MultiNavigationPolicy into .cpp file
